Fix user invite disabling
Since subject, recipient_user, and author_user have moved to
conversation, the code inside the user model to disable invites for a
user had to be modified. I can't fully remove it for now because I need
those three fields to convert messages to conversations. There is a
TODO to remove the extra code.