Start a new conversation
Users fill out a "To" field with a user's username, a "Subject", and a
"Message" body.

The "value" attribute here on the username field is an unfortunate
hack. @user is a variable that holds the currently logged in user. The
form helpers pick up the instance variable and fill in the attribute of
username. However, no one is trying to send a message to the currently
logged in user (themselves).

The actual creation of the conversation creates not only a
conversation, but also a message. It creates the message only after it
knows the conversation is valid and created. It also must fetch the
user to see that username matches anyone in the system. If not it will
return an error.

The result is a bit of a message in the controller. This calls for some
kind of object that creates conversations and takes these edge cases
into account. However, now we have a working test that we can refactor
against.