Convert deleted messages to conversations
Conversation deletion is odd when there are two people with access to
the conversation. What happens if the author deletes it and the
recipient sends a new message afterward?

We decided that the conversation would store an
author/recipient_deleted_at timestamp. If a user deletes the
conversation, it deletes set the appropriate column to the current time.

Then, if a message is added, it will update the timestamp on the
conversation table and the conversation will reappear for the user who
deleted it. This works because we can compare the updated_at timestamp
to the author/recipient_deleted_at. If updated_at is after _deleted_at,
we will display the conversation.